+++ src/Composer/ProcessHelper.php
@@ -251,6 +251,7 @@ final class ProcessHelper
      */
     private function getPackageRootDirectoryPath(PackageInterface $package): string
     {
+        /** @var string Because we use library and composer-plugins only, which always has installation path. */
         return $this->composer
             ->getInstallationManager()
             ->getInstallPath($package);